+++ CHANGELOG.md
@@ -32,6 +32,7 @@
 - (Feature) (Platform) Package Chart Generation
 - (Feature) (DebugPackage) CrashHandler Data
 - (Feature) (DebugPackage) Activities
+- (Feature) (Platform) Package Copy CLI
 
 ## [1.4.1](https://github.com/arangodb/kube-arangodb/tree/1.4.1) (2026-01-27)
 - (Bugfix) (Platform) Ensure Helm History limit in ArangoPlatformService